# api/openapi.yaml
openapi: 3.0.3
info:
  title: Counterline Terminal Access API
  version: 1.0.0
  description: |
    Terminal access only. Owner email authentication remains with Supabase.
    Serve through a same-origin /api proxy. Mutations require the exact configured
    Origin. Responses are no-store. Credentials are opaque HttpOnly, SameSite=Strict
    cookies (Secure in production), never JSON or browser storage tokens.
    Device access lasts 15 minutes. Device refresh sessions last 30 days and retain
    a 60-second single-use retry window after rotation for lost-response recovery.
    Refresh revalidates an unexpired cashier session; changed/inactive employees lock.
    Provision and refresh return only the active employee credential projection.
    Offline access is limited to seven days; manager approvals to 72 hours.
    Cashier catalog bootstrap and checkout use the terminal access plus cashier-session
    cookies. They are scoped to the terminal's store; a caller cannot choose another store.
servers:
  - url: /api
security: []
paths:
...
  /reports/daily-summary:
    get:
      operationId: getDailySummary
      summary: Aggregate a store's sales for one calendar day, computed in the store's own timezone
      security: [{ ownerBearer: [] }]
      parameters:
        - { name: store_id, in: query, required: true, schema: { type: string, format: uuid } }
        - { name: date, in: query, required: true, schema: { type: string, format: date }, description: Calendar date (YYYY-MM-DD) in the store's own timezone. }
      responses:
        '200': { description: Store-wide totals across every device, content: { application/json: { schema: { $ref: '#/components/schemas/DailySummary' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /reports/orders:
    get:
      operationId: getReportOrders
      summary: Cross-device drill-down of a store's orders for one calendar day
      security: [{ ownerBearer: [] }]
      parameters:
        - { name: store_id, in: query, required: true, schema: { type: string, format: uuid } }
        - { name: date, in: query, required: true, schema: { type: string, format: date } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        '200': { description: Store-scoped order page and next cursor, content: { application/json: { schema: { $ref: '#/components/schemas/ReportOrdersPage' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /reports/oversold:
    get:
      operationId: getOversold
      summary: List products with negative server-truth stock for a store
      security: [{ ownerBearer: [] }]
      parameters:
        - { name: store_id, in: query, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Products whose pos_stock.current_stock is below zero, most oversold first
          content:
            application/json:
              schema:
                type: object
                required: [products]
                properties:
                  products: { type: array, items: { $ref: '#/components/schemas/OversoldProduct' } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    ownerBearer: { type: http, scheme: bearer, bearerFormat: Supabase JWT }
    deviceAccess: { type: apiKey, in: cookie, name: terminal_access }
    deviceRefresh: { type: apiKey, in: cookie, name: terminal_refresh }
    cashierSession: { type: apiKey, in: cookie, name: terminal_cashier }
  responses:
    Error:
      description: '400 invalid request; 401 invalid credentials/PIN; 403 origin or manager access denied; 404 entity missing in store; 429 lockout; 500 unavailable. HTTP authentication failures must not fall back to cached login.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    DailySummary:
      type: object
      required: [grossSalesCents, discountCents, netSalesCents, taxCents, cashTakingsCents, cardTakingsCents, recordedTotalCents, completedOrderCount, averageSaleCents, itemsSold, refundedCount, refundedAmountCents]
      description: Field names match LocalSalesReport in apps/web/src/lib/reporting.ts so the frontend can consume either shape uniformly.
      properties:
        grossSalesCents: { type: integer }
        discountCents: { type: integer }
        netSalesCents: { type: integer }
        taxCents: { type: integer }
        cashTakingsCents: { type: integer }
        cardTakingsCents: { type: integer }
        recordedTotalCents: { type: integer }
        completedOrderCount: { type: integer }
        averageSaleCents: { type: integer }
        itemsSold: { type: integer }
        refundedCount: { type: integer }
        refundedAmountCents: { type: integer }
    ReportOrderSummary:
      type: object
      required: [id, receiptNumber, time, totalCents, paymentMethod, itemCount, syncStatus, employeeId, cashierName, refunded]
      properties:
        id: { type: string, format: uuid }
        receiptNumber: { type: string }
        time: { type: string, format: date-time }
        totalCents: { type: integer }
        paymentMethod: { type: string, enum: [cash, card, unknown] }
        itemCount: { type: integer }
        syncStatus: { type: string, enum: [synced] }
        employeeId: { type: string, format: uuid, nullable: true }
        cashierName: { type: string, nullable: true }
        refunded: { type: boolean }
    ReportOrdersPage:
      type: object
      required: [orders, next_cursor]
      properties:
        orders: { type: array, items: { $ref: '#/components/schemas/ReportOrderSummary' } }
        next_cursor: { type: string, nullable: true, description: Opaque pagination cursor; null means no further page. }
    OversoldProduct:
      type: object
      required: [id, name, sku, current_stock]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        sku: { type: string }
        current_stock: { type: integer, description: Negative when oversold. }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, enum: [validation_failed, authentication_required, origin_rejected, manager_required, employee_not_found, device_not_found, pin_locked, pin_invalid, server_unavailable] }
        message: { type: string }
...